pyproject.toml
[tool.poetry]
name = "kvali"
version = "0.1.0"
description = ""
authors = ["drain678 <ekaterina1.zorina@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^6.0.1"
djangorestframework = "^3.16.1"
psycopg2-binary = "^2.9.11"
python-dotenv = "^1.2.1"
django-prometheus = "^2.4.1"
prometheus-client = "^0.24.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.ruff]
target-version = "py312"
line-length = 88
exclude = [
    ".git",
    ".venv",
    "__pycache__",
    "migrations",
]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "F",   # pyflakes
    "I",   # isort
    "B",   # bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]

ignore = [
    "E501",  # line too long (если не используешь ruff format)
]

fixable = ["ALL"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "lf"
